rename pkg to 'unlock'
